chore: Remove the GitHub update checker
Update notifications will now come through the flashlight prism-notices
endpoint instead of a daily GitHub releases poll. This drops:

- UpdateCheckerThread (and its scheduling in start_threads / on settings save)
- the update_available_event on the controller and the related info_cell in
  the overlay
- the GitHub HTTP fetch (get_latest_release_tag, parse_releases_to_latest_tag)
  and the top-level update_available helper

VersionInfo (and the parse / update_available method on it) is still used to
track greatest_version in settings, so update_checker.py is renamed to
version.py with only the parsing/comparison primitives kept.
