+++ tests/prism/overlay/utils.py
@@ -407,7 +407,6 @@ def create_controller(
     autowho_event_set: bool = False,
     redraw_event_set: bool = False,
     game_ended_event_set: bool = False,
-    update_available_event_set: bool = False,
     current_player_updates: Iterable[KnownPlayer] | None = None,
     player_cache: PlayerCache | None = None,
     nick_database: NickDatabase | None = None,
@@ -437,8 +436,6 @@ def create_controller(
         controller.redraw_event.set()
     if game_ended_event_set:
         controller.game_ended_event.set()
-    if update_available_event_set:
-        controller.update_available_event.set()
 
     if player_cache is not None:
         controller.player_cache = player_cache
@@ -485,10 +482,6 @@ def assert_controllers_equal(
     assert controller1.redraw_event.is_set() == controller2.redraw_event.is_set()
     assert (
         controller1.game_ended_event.is_set() == controller2.game_ended_event.is_set()
-    )
-    assert (
-        controller1.update_available_event.is_set()
-        == controller2.update_available_event.is_set()
     )
 
     if not ignore_player_cache:
